preparando coisas para a integracao com o banco

+++ src/main/java/com/nossogrupo/GerenciadorTarefas/repository/TarefaRepository.java
@@ -3,6 +3,6 @@ package com.nossogrupo.GerenciadorTarefas.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.nossogrupo.GerenciadorTarefas.model.Tarefa;
 
-public interface TarefaRepository extends JpaRepository<Tarefa, Long>{
+public interface TarefaRepository extends JpaRepository<Tarefa, Integer>{
     
 }
