+++ src/main/java/com/nossogrupo/GerenciadorTarefas/repository/TaskUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nossogrupo.GerenciadorTarefas.model.TaskUser;
 
-public interface TaskUserRepository extends JpaRepository<TaskUser, Long>{
+public interface TaskUserRepository extends JpaRepository<TaskUser, Integer>{
     
 }
